clippy fixes

+++ kzg/src/common_utils.rs
@@ -65,7 +65,7 @@ pub fn log_2(x: usize) -> usize {
     if x == 0 {
         return 0;
     }
-    num_bits::<usize>() as usize - (x.leading_zeros() as usize) - 1
+    num_bits::<usize>() - (x.leading_zeros() as usize) - 1
 }
 
 pub fn is_power_of_2(n: usize) -> bool {
